+++ src/commands/self_cmd/cli.rs
@@ -12,6 +12,7 @@ pub struct Cli {
 #[derive(Subcommand, Debug)]
 enum Commands {
     Metadata(super::metadata::Cli),
+    Pip(super::pip::Cli),
     Python(super::python::Cli),
     Restore(super::restore::Cli),
     Update(super::update::Cli),
@@ -21,6 +22,7 @@ impl Cli {
     pub fn exec(self) -> Result<()> {
         match self.command {
             Commands::Metadata(cli) => cli.exec(),
+            Commands::Pip(cli) => cli.exec(),
             Commands::Python(cli) => cli.exec(),
             Commands::Restore(cli) => cli.exec(),
             Commands::Update(cli) => cli.exec(),